Move database sorting to same sortingHelpers file with variable sorting
